refactor: hypertrons
