+++ src/pages/ContentScripts/index.ts
@@ -6,6 +6,7 @@ import PerceptorTab from '../../anchors/PerceptorTab';
 import PerceptorLayout from '../../anchors/PerceptorLayout';
 import DeveloperNetwork from '../../anchors/DeveloperNetwork';
 import ProjectNetwork from '../../anchors/ProjectNetwork';
+import Hypertrons from '../../anchors/Hypertrons';
 import { inject2Perceptor } from './Perceptor';
 initializeIcons();
 inject2Perceptor(DeveloperActiInflTrend);
@@ -15,7 +16,7 @@ inject2Perceptor(PerceptorTab);
 inject2Perceptor(PerceptorLayout);
 inject2Perceptor(DeveloperNetwork);
 inject2Perceptor(ProjectNetwork);
-import './Hypertrons';
+inject2Perceptor(Hypertrons);
 import './content.styles.css';
 import { Perceptor } from './Perceptor';
 import { loadSettings } from '../../utils/settings';
