+++ src/views/Hypertrons.tsx
@@ -1,8 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { render } from 'react-dom';
-import $ from 'jquery';
 import { fire } from 'delegated-events';
-import * as pageDetect from 'github-url-detection';
 import {
   Callout,
   Stack,
@@ -13,19 +10,14 @@ import {
   DirectionalHint,
 } from '@fluentui/react';
 import { useBoolean } from '@fluentui/react-hooks';
-import { utils } from 'github-url-detection';
 import {
   Command,
   LabelStyles,
   Label2Style,
   getUserNameFromCookie,
-} from '../../services/hypertrons';
-import { getMessageByLocale, runsWhen } from '../../utils/utils';
-import PerceptorBase from './PerceptorBase';
-import { inject2Perceptor } from './Perceptor';
-import logger from '../../utils/logger';
-import Settings, { loadSettings } from '../../utils/settings';
-import { getConfigFromGithub } from '../../api/github';
+} from '../services/hypertrons';
+import { getMessageByLocale, runsWhen } from '../utils/utils';
+import Settings, { loadSettings } from '../utils/settings';
 
 const styles = mergeStyleSets({
   callout: {
@@ -224,40 +216,4 @@ const HypertronsTabView: React.FC<HypertronsTabViewProps> = ({
   );
 };
 
-@runsWhen([pageDetect.isPR, pageDetect.isIssue])
-class Hypertrons extends PerceptorBase {
-  public static hypertronsConfig: any;
-
-  private static renderView(): void {
-    // avoid redundant button
-    if ($('#hypertrons_button').length > 0) {
-      logger.info('hypertrons tab exists');
-      return;
-    }
-
-    // add hypertrons tab
-    const commentForm = $('.js-new-comment-form');
-    const parentContainer = commentForm.find('.d-flex.flex-justify-end');
-    const hypertronsTab = document.createElement('div');
-    render(
-      <HypertronsTabView hypertronsConfig={this.hypertronsConfig} />,
-      hypertronsTab
-    );
-    parentContainer.prepend(hypertronsTab);
-  }
-
-  public async run(config: any): Promise<void> {
-    Hypertrons.hypertronsConfig = config;
-    // @ts-ignore
-    const observer = new MutationObserver(Hypertrons.renderView);
-    const element = document.querySelector('#new_comment_field');
-    // @ts-ignore
-    observer.observe(element, {
-      attributes: true,
-    });
-
-    Hypertrons.renderView();
-  }
-}
-
-inject2Perceptor(Hypertrons);
+export default HypertronsTabView;
